+++ etl/penetrations/source.rb
@@ -11,10 +11,10 @@ module Penetrations
       super { sheets.select { |row| row[PenetrationsImporter::ZIP].presence }.each(&block) }
     end
 
-    # This method removes the array with nil elements and removes leading and trailing whitespaces
-    # in the string. Then the integer value is inserted back again and formatted array is returned.
+    # This method removes the array elements of type string with leading and trailing whitespaces
+    # Then the integer value is inserted back again and formatted array is returned.
     def format(value)
-      value.compact.map do |index|
+      value.map do |index|
         index.instance_of?(String) ? index.strip : index
       end
     end