+++ spec/graphql/mutations/create_project_spec.rb
@@ -21,6 +21,7 @@ RSpec.describe Mutations::CreateProject do
       :hfc_footprint,
       zip: '8008',
       kam_region: kam_region,
+      strategic_partner: :cable_group_ag,
       penetration_competitions: [build(:penetration_competition, competition: create(:admin_toolkit_competition))]
     )
   end
@@ -206,6 +207,7 @@ RSpec.describe Mutations::CreateProject do
               lotNumber: "EA0988833"
               buildingsCount: 1
               buildingType: "efh"
+              moveInStartsOn: "2021-12-12"
               #{address}
               #{address_books}
               #{site_address}
@@ -215,12 +217,13 @@ RSpec.describe Mutations::CreateProject do
         {
           project {
             id status category internalId assigneeType lotNumber
-            buildingsCount buildingType
+            buildingsCount buildingType moveInStartsOn
             address { city id street streetNo zip }
             addressBooks { id type name company language email website phone mobile address {#{' '}
               city id street streetNo zip } }
             assignee { id name }
             buildingType
+            strategicPartner
             kamRegion { id name }
             siteAddress
           }